Move titlebar things into title_bar
