+++ crates/storybook/src/story_selector.rs
@@ -35,7 +35,7 @@ pub enum ComponentStory {
     Tab,
     TabBar,
     Text,
-    TitleBar,
+    // TitleBar,
     ToggleButton,
     ToolStrip,
     ViewportUnits,
@@ -69,7 +69,7 @@ impl ComponentStory {
             Self::Text => TextStory::view(cx).into(),
             Self::Tab => cx.new_view(|_| ui::TabStory).into(),
             Self::TabBar => cx.new_view(|_| ui::TabBarStory).into(),
-            Self::TitleBar => cx.new_view(|_| ui::TitleBarStory).into(),
+            // Self::TitleBar => cx.new_view(|_| title_bar::TitleBarStory).into(),
             Self::ToggleButton => cx.new_view(|_| ui::ToggleButtonStory).into(),
             Self::ToolStrip => cx.new_view(|_| ui::ToolStripStory).into(),
             Self::ViewportUnits => cx.new_view(|_| crate::stories::ViewportUnitsStory).into(),
