+++ crates/title_bar/Cargo.toml
@@ -74,6 +74,9 @@ vcs_menu.workspace = true
 workspace.workspace = true
 zed_actions.workspace = true
 
+[target.'cfg(windows)'.dependencies]
+windows.workspace = true
+
 [dev-dependencies]
 call = { workspace = true, features = ["test-support"] }
 client = { workspace = true, features = ["test-support"] }
