+++ crates/title_bar/src/platforms.rs
@@ -1,0 +1,3 @@
+pub mod linux;
+pub mod mac;
+pub mod windows;
