+++ crates/ui/src/components.rs
@@ -20,7 +20,6 @@ mod setting;
 mod stack;
 mod tab;
 mod tab_bar;
-mod title_bar;
 mod tool_strip;
 mod tooltip;
 
@@ -49,7 +48,6 @@ pub use setting::*;
 pub use stack::*;
 pub use tab::*;
 pub use tab_bar::*;
-pub use title_bar::*;
 pub use tool_strip::*;
 pub use tooltip::*;
