+++ crates/ui/src/components/stories.rs
@@ -13,7 +13,6 @@ mod list_item;
 mod setting;
 mod tab;
 mod tab_bar;
-mod title_bar;
 mod toggle_button;
 mod tool_strip;
 
@@ -32,6 +31,5 @@ pub use list_item::*;
 pub use setting::*;
 pub use tab::*;
 pub use tab_bar::*;
-pub use title_bar::*;
 pub use toggle_button::*;
 pub use tool_strip::*;
